Suggested minor tweaks to CDR Tutorials 1 & 2

+++ snorkel/utils.py
@@ -12,8 +12,8 @@ class ProgressBar(object):
 
     def bar(self, i):
         """Assumes i ranges through [0, N-1]"""
-        b = np.ceil(((i+1) / self.nf) * self.length)
-        sys.stdout.write("\r[%s%s] %d%%" % ("="*b, " "*(self.length-b), 100*((i+1) / self.nf)))
+        b = int(np.ceil(((i+1) / self.nf) * self.length))
+        sys.stdout.write("\r[%s%s] %d%%" % ("="*b, " "*(self.length-b), int(100*((i+1) / self.nf))))
         sys.stdout.flush()
 
     def close(self):
